[audio] Implement audio filters for clay web too! (still need to handle params)

+++ plugins/web/tools/src/tools/tasks/web/Minify.hx
@@ -20,39 +20,43 @@ class Minify extends tools.Task {
     override function run(cwd:String, args:Array<String>):Void {
 
         var project = ensureCeramicProject(cwd, args, App);
-        var jsName:String = project.app.name;
+        var jsNames:Array<String> = [project.app.name, 'audio-worklets'];
 
         var webProjectPath = Path.join([cwd, 'project/web']);
 
-        var cmdArgs = [];
-        cmdArgs.push(Path.join([context.plugins.get('web').path, 'resources/uglifyjs/node_modules/.bin/uglifyjs']));
-        cmdArgs.push('$jsName.js');
+        for (jsName in jsNames) {
+            if (FileSystem.exists(Path.join([webProjectPath, '$jsName.js']))) {
+                var cmdArgs = [];
+                cmdArgs.push(Path.join([context.plugins.get('web').path, 'resources/uglifyjs/node_modules/.bin/uglifyjs']));
+                cmdArgs.push('$jsName.js');
 
-        print('Minify javascript');
+                print('Minify javascript: $jsName');
 
-        if (context.debug || context.defines.exists('js-source-map')) {
-            if (FileSystem.exists(Path.join([webProjectPath, '$jsName.js.map']))) {
-                cmdArgs.push('--source-map');
-                cmdArgs.push('content=\'$jsName.js.map\'');
+                if (context.debug || context.defines.exists('js-source-map')) {
+                    if (FileSystem.exists(Path.join([webProjectPath, '$jsName.js.map']))) {
+                        cmdArgs.push('--source-map');
+                        cmdArgs.push('content=\'$jsName.js.map\'');
+                    }
+                }
+                else {
+                    var minSourceMap = Path.join([webProjectPath, '$jsName.min.js.map']);
+                    if (FileSystem.exists(minSourceMap)) {
+                        FileSystem.deleteFile(minSourceMap);
+                    }
+                }
+
+                cmdArgs.push('--compress');
+                cmdArgs.push('--mangle');
+                cmdArgs.push('--keep-fnames');
+
+                cmdArgs.push('--output');
+                cmdArgs.push('$jsName.min.js');
+
+                node(cmdArgs, {
+                    cwd: webProjectPath
+                });
             }
         }
-        else {
-            var minSourceMap = Path.join([webProjectPath, '$jsName.min.js.map']);
-            if (FileSystem.exists(minSourceMap)) {
-                FileSystem.deleteFile(minSourceMap);
-            }
-        }
-
-        cmdArgs.push('--compress');
-        cmdArgs.push('--mangle');
-        cmdArgs.push('--keep-fnames');
-
-        cmdArgs.push('--output');
-        cmdArgs.push('$jsName.min.js');
-
-        node(cmdArgs, {
-            cwd: webProjectPath
-        });
 
     }
